LayoutTests/editing/inserting/insert-composition-whitespace-expected.txt

PASS compositingText is 'AB'
PASS compositingText is 'A B'
PASS compositingText is 'A  B'
PASS compositingText is 'A   B'
PASS compositingText is 'A    B'
PASS compositingText is ' AB'
PASS compositingText is '  AB'
PASS compositingText is '   AB'
PASS compositingText is '    AB'
PASS compositingText is '     AB'
PASS compositingText is '      AB'
PASS compositingText is '       AB'
PASS compositingText is 'AB  '
PASS compositingText is 'AB   '
PASS compositingText is 'AB    '
PASS compositingText is 'AB     '
PASS compositingText is 'AB      '
PASS compositingText is 'AB       '
PASS compositingText is '  A  B  '
PASS compositingText is '  A  B  '
PASS compositingText is ' '
PASS compositingText is '  '
PASS compositingText is '   '
PASS compositingText is 'AB'
PASS compositingText is 'A B'
PASS compositingText is 'A  B'
PASS compositingText is 'A   B'
PASS compositingText is 'A    B'
PASS compositingText is 'AB  '
PASS compositingText is 'AB   '
PASS compositingText is 'AB    '
PASS compositingText is 'AB     '
PASS compositingText is 'AB      '
PASS compositingText is 'AB       '
PASS compositingText is '  AB'
PASS compositingText is '   AB'
PASS compositingText is '    AB'
PASS compositingText is '     AB'
PASS compositingText is '      AB'
PASS compositingText is '       AB'
PASS compositingText is '  A  B  '
PASS compositingText is '		A		B		'
PASS compositingText is ' '
PASS compositingText is '  '
PASS compositingText is '   '
PASS successfullyParsed is true

TEST COMPLETE
